requirements.txt
flask
pandas
openpyxl
psycopg2-binary
gunicorn